resolve conversation

+++ src/main/java/org/gooru/dap/constants/Constants.java
@@ -14,6 +14,6 @@ public final class Constants {
   public static final String CORE_DATA_SOURCE = "coreDataSource";
   public static final String REPORTS_DATA_SOURCE = "reportsDataSource";
   public static final double MASTERY_SCORE = 80.00;
-  public static final int DEFAULT_COMPLETED_SCORE = 80;
+  public static final double DEFAULT_COMPLETED_SCORE = 80.00;
 
 }
